Linked CAF Buttons to CAF Canvas
Added script to CAF Canvas and linked buttons to CAF Canvas. Now onclick for each button will execute the script in the CAF Canvas instead of in the indivudual button.

Note: This is needed to store global variables as part of the button's parent (the CAF canvas) instead of as part of the button itself.